feat(ui): updated display of the number of products in the cart in the Header component, implemented actions for adding, subtracting, and deleting products in the BasketPage component, added information about the total price of products in the cart
In this commit, the display of the number of products in the cart within the Header component has been updated. Actions for adding, subtracting, and deleting individual products have also been implemented in the BasketPage component. Information about the total price of all products in the cart has been added to allow users to see the overall cost of their cart.